[SC-6147]: Remove empty folders after upgrade

+++ sugarcrm/upgrade/scripts/post/9_RemoveFiles.php
@@ -28,7 +28,7 @@ class SugarUpgradeRemoveFiles extends UpgradeScript
         }
 
         $caseInsensitiveFS = $this->upgrader->context['case_insensitive_fs'];
-        $folders = array();
+        $foldersToCheck = array();
 
         foreach ($this->state['files_to_delete'] as $file) {
             $file = SugarAutoLoader::normalizeFilePath($file);
@@ -43,17 +43,24 @@ class SugarUpgradeRemoveFiles extends UpgradeScript
 
             if (is_dir($file)) {
                 $this->removeDir($file);
+                $foldersToCheck[$file] = false;
             } else {
                 $this->unlink($file);
                 $folder = dirname($file);
 
-                if (!isset($folders[$folder])) {
-                    $folders[$folder] = count(glob($folder . '/*'));
+                if (!isset($foldersToCheck[$folder])) {
+                    $foldersToCheck[$folder] = true;
                 }
             }
         }
 
-        foreach ($folders as $folder => $numberOfFiles) {
+        foreach ($foldersToCheck as $folder => $remove) {
+            if (!$remove) {
+                continue;
+            }
+
+            $numberOfFiles = count(glob($folder . '/*'));
+
             if ($numberOfFiles === 0) {
                 $this->removeDir($folder);
             }
